images table, show post.html, save likes count

+++ src/main/java/myblog/model/Comment.java
@@ -1,16 +1,16 @@
 package myblog.model;
 
 public class Comment {
-  private int id;
+  private long id;
   private String text;
 
-  public Comment( int id, String text ){
+  public Comment( long id, String text ){
     this.id = id;
     this.text = text;
   }
-  public void setId( int id ) { this.id = id; }
+  public void setId( long id ) { this.id = id; }
   public void setText( String text ) { this.text = text; }
-  public int getId() { return id; }
+  public long getId() { return id; }
   public String getText() { return text; }
 
 }
